// frontend/src/types/basketballStats.ts
export type BasketballStats = {
  player_name: string;
  team: string;
  conf: string;
  GP: number;
  Min_per: number;
  ORtg: number;
  usg: number;
  eFG: number;
  TS_per: number;
  ORB_per: number;
  DRB_per: number;
  AST_per: number;
  TO_per: number;
  FTM: number;
  FTA: number;
  FT_per: number;
  twoPM: number;
  twoPA: number;
  twoP_per: number;
  TPM: number;
  TPA: number;
  TP_per: number;
  blk_per: number;
  stl_per: number;
  ftr: number;
  yr: string;
  ht: string;
  num: number;
  porpag: number;
  adjoe: number;
  pfr: number;
  year: number;
  pid: number;
  type: string;
  recRank: string;
  astTov: number;
  rimmade: number;
  rimmadeRimmiss: number;
  midmade: number;
  midmadeMidmiss: number;
  rimmadeRatio: number;
  midmadeRatio: number;
  dunksmade: number;
  dunksTotal: number;
  dunksRatio: number;
  pick: number;
  drtg: number;
  adrtg: number;
  dporpag: number;
  stops: number;
  bpm: number;
  obpm: number;
  dbpm: number;
  gbmp: number;
  mp: number;
  ogbpm: number;
  dgbpm: number;
  oreb: number;
  dreb: number;
  treb: number;
  ast: number;
  stl: number;
  blk: number;
  pts: number;
  role: string;
  threePPer100: number;
}; 